Simplify Xdebug and browser globals wiring
