+++ packages/php-wasm/web/src/test/php-dynamic-loading.spec.ts
@@ -19,7 +19,7 @@ SupportedPHPVersions.forEach((phpVersion) => {
 
 			await page.addScriptTag({
 				type: 'module',
-				url: '/src/test/playwright/globals.ts',
+				url: '/src/test/playwright/browser-globals.ts',
 			});
 		});
