+++ packages/php-wasm/web/src/test/php-memory-deallocation.spec.ts
@@ -21,7 +21,7 @@ test.describe(`Memory Deallocation - PHP ${LatestSupportedPHPVersion}`, () => {
 
 		await page.addScriptTag({
 			type: 'module',
-			url: '/src/test/playwright/globals.ts',
+			url: '/src/test/playwright/browser-globals.ts',
 		});
 	});
